+++ src/lib/util.js
@@ -258,3 +258,19 @@ function addParamsToUrl (url, object, encodeComponents) {
 }
 
 module.exports.addParamsToUrl = addParamsToUrl;
+
+/**
+ * isDomAvailable
+ * @description Checks to see if the DOM is available by checking the existence of the window and document
+ * @see https://github.com/facebook/fbjs/blob/master/packages/fbjs/src/core/ExecutionEnvironment.js#L12
+ */
+
+function isDomAvailable () {
+  return (
+    typeof window !== 'undefined' &&
+    !!window.document &&
+    !!window.document.createElement
+  );
+}
+
+module.exports.isDomAvailable = isDomAvailable;